pyproject.toml
[tool.poetry]
name = "cartierville"
version = "0.1.0"
description = "Radio stream to podcast"
authors = ["Donald AC Martin <donald@donald-martin.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.26"
schedule = "^1.1.0"
PyYAML = "^6.0"
Cerberus = "^1.3.4"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
pylint = "^2.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
